LayoutTests/imported/w3c/web-platform-tests/wasm/serialization/module/broadcastchannel-success-and-failure-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: DataCloneError: The object can not be cloned.

Harness Error (FAIL), message = Unhandled rejection: The object can not be cloned.

TIMEOUT WebAssembly.Module cannot cross agent clusters, BroadcastChannel edition Test timed out

Harness Error (FAIL), message = Unhandled rejection: The object can not be cloned.

TIMEOUT WebAssembly.Module cannot cross agent clusters, BroadcastChannel edition Test timed out
